pyproject.toml
[project]
name = "nextalk"
version = "0.1.0"
description = "个人轻量级实时语音识别与输入系统"
authors = [{name = "NexTalk Team", email = "contact@nextalk.dev"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "websockets>=12.0",
    "PyYAML>=6.0",
    "pynput>=1.7.6",
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "sounddevice>=0.4.6",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
gui = [
    "PyQt6>=6.4.0",
    "pillow>=9.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "coverage>=7.0.0",
]

[project.scripts]
nextalk = "nextalk.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.coverage.run]
source = ["nextalk"]
omit = ["*/tests/*"]
